requirement.txt
Django
django-filter
djangorestframework
djangorestframework-simplejwt
pytz
django-rest-swagger
gunicorn